remove parallel routes for browse
