--- app/browse/@results/loading.tsx
@@ -1,5 +1,0 @@
-import LoadingSkeleton from "@/components/LoadingSkeleton"
-
-export default function Loading() {
-  return <LoadingSkeleton />
-}
