+++ components/LoadingSkeleton.tsx
@@ -8,7 +8,7 @@ const dmSerifDisplay = DM_Serif_Display({ weight: "400", subsets: ["latin"] })
 const LoadingSkeleton: FC = () => {
   return (
     <PageContainer className="relative items-center justify-center">
-      <div className="animate-theme-slow-spin absolute h-40 w-40 rounded-full border-8 border-dotted border-slate-300"></div>
+      <div className="animate-theme-slow-spin absolute h-40 w-40 rounded-full border-8 border-dotted border-slate-300" />
       <h1
         className={twMerge(
           dmSerifDisplay.className,
